refactor(demo): move demo endpoints + SQLite into DemoModule (ADR-0001)
Root AppController/AppService keep only IdP concerns (health, public,
hydra accept). Demo endpoints, DemoService and the SQLite/TypeORM root
move into DemoModule; synchronize is now gated off in production.
